Move control flow functions to ControlFlow class
Control flow functions were divided over BasicLogic, DataHandling and Echoes. This commit creates one class with all control flow (branching) functions.
- Move control flow functions to ControlFlow.
- Move corresponding test cases to ControlFlowTest.
- Fix 120 character line limit in moved code.
