+++ src/main/java/com/laytonsmith/core/functions/BasicLogic.java
@@ -2,7 +2,6 @@ package com.laytonsmith.core.functions;
 
 import com.laytonsmith.PureUtilities.Version;
 import com.laytonsmith.annotations.api;
-import com.laytonsmith.annotations.breakable;
 import com.laytonsmith.annotations.core;
 import com.laytonsmith.annotations.seealso;
 import com.laytonsmith.core.ArgumentValidation;
@@ -11,23 +10,15 @@ import com.laytonsmith.core.Optimizable;
 import com.laytonsmith.core.ParseTree;
 import com.laytonsmith.core.Script;
 import com.laytonsmith.core.Static;
-import com.laytonsmith.core.compiler.BranchStatement;
 import com.laytonsmith.core.compiler.FileOptions;
 import com.laytonsmith.core.compiler.OptimizationUtilities;
-import com.laytonsmith.core.compiler.VariableScope;
 import com.laytonsmith.core.constructs.CArray;
 import com.laytonsmith.core.constructs.CBoolean;
-import com.laytonsmith.core.constructs.CFunction;
-import com.laytonsmith.core.constructs.CIdentifier;
 import com.laytonsmith.core.constructs.CInt;
-import com.laytonsmith.core.constructs.CKeyword;
-import com.laytonsmith.core.constructs.CLabel;
 import com.laytonsmith.core.constructs.CNull;
-import com.laytonsmith.core.constructs.CSlice;
 import com.laytonsmith.core.constructs.CString;
 import com.laytonsmith.core.constructs.CSymbol;
 import com.laytonsmith.core.constructs.CVoid;
-import com.laytonsmith.core.constructs.Construct;
 import com.laytonsmith.core.constructs.Target;
 import com.laytonsmith.core.environments.Environment;
 import com.laytonsmith.core.environments.GlobalEnv;
@@ -38,14 +29,11 @@ import com.laytonsmith.core.exceptions.CRE.CREThrowable;
 import com.laytonsmith.core.exceptions.CancelCommandException;
 import com.laytonsmith.core.exceptions.ConfigCompileException;
 import com.laytonsmith.core.exceptions.ConfigRuntimeException;
-import com.laytonsmith.core.exceptions.LoopBreakException;
 import com.laytonsmith.core.natives.interfaces.Mixed;
-import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
-import java.util.TreeSet;
 
 /**
  *
@@ -55,745 +43,6 @@ public class BasicLogic {
 
 	public static String docs() {
 		return "These functions provide basic logical operations.";
-	}
-
-	@api
-	public static class _if extends AbstractFunction implements Optimizable, BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "if";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{Integer.MAX_VALUE};
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			for(ParseTree node : nodes) {
-				if(node.getData() instanceof CIdentifier) {
-					return new ifelse().execs(t, env, parent, nodes);
-				}
-			}
-			ParseTree condition = nodes[0];
-			ParseTree __if = nodes[1];
-			ParseTree __else = null;
-			if(nodes.length == 3) {
-				__else = nodes[2];
-			}
-
-			if(ArgumentValidation.getBoolean(parent.seval(condition, env), t)) {
-				return parent.seval(__if, env);
-			} else {
-				if(__else == null) {
-					return CVoid.VOID;
-				}
-				return parent.seval(__else, env);
-			}
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws CancelCommandException, ConfigRuntimeException {
-			return CVoid.VOID;
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CRECastException.class};
-		}
-
-		@Override
-		public String docs() {
-			return "mixed {cond, trueRet, [falseRet]} If the first argument evaluates to a true value, the second argument is returned, otherwise the third argument is returned."
-					+ " If there is no third argument, it returns void.";
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public boolean preResolveVariables() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_0_1;
-		}
-		//Doesn't matter, this function is run out of state
-
-		@Override
-		public Boolean runAsync() {
-			return false;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(
-					OptimizationOption.OPTIMIZE_DYNAMIC
-			);
-		}
-
-		@SuppressWarnings("checkstyle:constantname")
-		private static final String and = new and().getName();
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> args, FileOptions fileOptions) throws ConfigCompileException {
-			//Check for too many/few arguments
-			if(args.size() < 2) {
-				throw new ConfigCompileException("Too few arguments passed to if()", t);
-			}
-			if(args.size() > 3) {
-				throw new ConfigCompileException("if() can only have 3 parameters", t);
-			}
-			if(args.get(0).isConst()) {
-				// We can optimize this one way or the other, since the condition is const
-				if(ArgumentValidation.getBoolean(args.get(0).getData(), t)) {
-					// It's true, return the true condition
-					return args.get(1);
-				} else // If there are three args, return the else condition, otherwise,
-				// have it entirely remove us from the parse tree.
-				if(args.size() == 3) {
-					return args.get(2);
-				} else {
-					return Optimizable.REMOVE_ME;
-				}
-			}
-			// If the code looks like this:
-			// if(@a){
-			//		if(@b){
-			//		}
-			// }
-			// then we can turn this into if(@a && @b){ }, as they are functionally
-			// equivalent, and this construct tends to be faster (less stack frames, presumably).
-			// The caveat is that if the inner if statement has an else statement (or is ifelse)
-			// or there are other nodes inside the statement, or we have an else clause
-			// we cannot do this optimization, as it then has side effects.
-			if(args.get(1).getData() instanceof CFunction && args.get(1).getData().val().equals("if") && args.size() == 2) {
-				ParseTree _if = args.get(1);
-				if(_if.getChildren().size() == 2) {
-					// All the conditions are met, move this up
-					ParseTree myCondition = args.get(0);
-					ParseTree theirCondition = _if.getChildAt(0);
-					ParseTree theirCode = _if.getChildAt(1);
-					ParseTree andClause = new ParseTree(new CFunction(and, t), fileOptions);
-					// If it's already an and(), just tack the other condition on
-					if(myCondition.getData() instanceof CFunction && myCondition.getData().val().equals(and)) {
-						andClause = myCondition;
-						andClause.addChild(theirCondition);
-					} else {
-						andClause.addChild(myCondition);
-						andClause.addChild(theirCondition);
-					}
-					args.set(0, andClause);
-					args.set(1, theirCode);
-				}
-			}
-			return null;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Functional usage", "if(true, msg('This is true'), msg('This is false'))"),
-				new ExampleScript("With braces, true condition", "if(true){\n\tmsg('This is true')\n}"),
-				new ExampleScript("With braces, false condition", "msg('Start')\nif(false){\n\tmsg('This will not show')\n}\nmsg('Finish')")};
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> branches = new ArrayList<>(children.size());
-			// Only the first child is not a branch. Everything else is a branch.
-			branches.add(false);
-			for(int i = 1; i < children.size(); i++) {
-				branches.add(true);
-			}
-			return branches;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			// It's the exact same logic as the branches
-			return isBranch(children);
-		}
-
-	}
-
-	@api(environments = {GlobalEnv.class})
-	public static class ifelse extends AbstractFunction implements Optimizable, BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "ifelse";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{Integer.MAX_VALUE};
-		}
-
-		@Override
-		public String docs() {
-			return "mixed {[boolean1, code]..., [elseCode]} Provides a more convenient method"
-					+ " for running if/else chains. If none of the conditions are true, and"
-					+ " there is no 'else' condition, void is returned.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CREInsufficientArgumentsException.class};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public boolean preResolveVariables() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			return CNull.NULL;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			if(nodes.length < 2) {
-				throw new CREInsufficientArgumentsException("ifelse expects at least 2 arguments", t);
-			}
-			for(int i = 0; i <= nodes.length - 2; i += 2) {
-				ParseTree statement = nodes[i];
-				ParseTree code = nodes[i + 1];
-				Mixed evalStatement = parent.seval(statement, env);
-				if(evalStatement instanceof CIdentifier) {
-					evalStatement = parent.seval(((CIdentifier) evalStatement).contained(), env);
-				}
-				if(ArgumentValidation.getBoolean(evalStatement, t)) {
-					Mixed ret = env.getEnv(GlobalEnv.class).GetScript().eval(code, env);
-					return ret;
-				}
-			}
-			if(nodes.length % 2 == 1) {
-				Mixed ret = env.getEnv(GlobalEnv.class).GetScript().seval(nodes[nodes.length - 1], env);
-				if(ret instanceof CIdentifier) {
-					return parent.seval(((CIdentifier) ret).contained(), env);
-				} else {
-					return ret;
-				}
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Set<Optimizable.OptimizationOption> optimizationOptions() {
-			return EnumSet.of(
-					Optimizable.OptimizationOption.OPTIMIZE_DYNAMIC
-			);
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			// TODO: Redo this optimization.
-			return null;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Functional usage", "ifelse(false, msg('This is false'), true, msg('This is true'))"),
-				new ExampleScript("With braces", "if(false){\n\tmsg('This is false')\n} else {\n\tmsg('This is true')\n}"),
-				new ExampleScript("With braces, with else if", "if(false){\n\tmsg('This will not show')\n} else if(false){\n"
-				+ "\n\tmsg('This will not show')\n} else {\n\tmsg('This will show')\n}")};
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> branches = new ArrayList<>(children.size());
-			// Only the first child is not a branch. Everything else is a branch.
-			branches.add(false);
-			for(int i = 1; i < children.size(); i++) {
-				branches.add(true);
-			}
-			return branches;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			// It's the exact same logic as the branches
-			return isBranch(children);
-		}
-	}
-
-	@api
-	@breakable
-	public static class _switch extends AbstractFunction implements Optimizable, BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "switch";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{Integer.MAX_VALUE};
-		}
-
-		@Override
-		public String docs() {
-			return "mixed {value, [equals, code]..., [defaultCode]} Provides a switch statement. If none of the conditions"
-					+ " match, and no default is provided, void is returned."
-					+ " See the documentation on [[CommandHelper/Logic|Logic]] for more information. ----"
-					+ " In addition, slices may be used to indicate ranges of integers that should trigger the specified"
-					+ " case. Slices embedded in an array are fine as well. Switch statements also support brace/case/default"
-					+ " syntax, as in most languages, althrough unlike most languages, fallthrough isn't supported. Breaking"
-					+ " with break() isn't required, but recommended. A number greater than 1 may be sent to break, and breaking"
-					+ " out of the switch will consume a \"break counter\" and the break will continue up the chain."
-					+ " If you do use break(), the return value of switch is ignored. See the examples for usage"
-					+ " of brace/case/default syntax, which is highly recommended.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CREInsufficientArgumentsException.class};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public boolean preResolveVariables() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			return CNull.NULL;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			Mixed value = parent.seval(nodes[0], env);
-			equals equals = new equals();
-			try {
-				for(int i = 1; i <= nodes.length - 2; i += 2) {
-					ParseTree statement = nodes[i];
-					ParseTree code = nodes[i + 1];
-					Mixed evalStatement = parent.seval(statement, env);
-					if(evalStatement instanceof CSlice) { //More specific subclass of array, we can do more optimal handling here
-						long rangeLeft = ((CSlice) evalStatement).getStart();
-						long rangeRight = ((CSlice) evalStatement).getFinish();
-						if(value.isInstanceOf(CInt.class)) {
-							long v = Static.getInt(value, t);
-							if((rangeLeft < rangeRight && v >= rangeLeft && v <= rangeRight)
-									|| (rangeLeft > rangeRight && v >= rangeRight && v <= rangeLeft)
-									|| (rangeLeft == rangeRight && v == rangeLeft)) {
-								return parent.seval(code, env);
-							}
-						}
-					} else if(evalStatement.isInstanceOf(CArray.class)) {
-						for(String index : ((CArray) evalStatement).stringKeySet()) {
-							Mixed inner = ((CArray) evalStatement).get(index, t);
-							if(inner instanceof CSlice) {
-								long rangeLeft = ((CSlice) inner).getStart();
-								long rangeRight = ((CSlice) inner).getFinish();
-								if(value.isInstanceOf(CInt.class)) {
-									long v = Static.getInt(value, t);
-									if((rangeLeft < rangeRight && v >= rangeLeft && v <= rangeRight)
-											|| (rangeLeft > rangeRight && v >= rangeRight && v <= rangeLeft)
-											|| (rangeLeft == rangeRight && v == rangeLeft)) {
-										return parent.seval(code, env);
-									}
-								}
-							} else if(equals.exec(t, env, value, inner).getBoolean()) {
-								return parent.seval(code, env);
-							}
-						}
-					} else if(equals.exec(t, env, value, evalStatement).getBoolean()) {
-						return parent.seval(code, env);
-					}
-				}
-				if(nodes.length % 2 == 0) {
-					return parent.seval(nodes[nodes.length - 1], env);
-				}
-			} catch (LoopBreakException ex) {
-				//Ignored, unless the value passed in is greater than 1, in which case
-				//we rethrow.
-				if(ex.getTimes() > 1) {
-					ex.setTimes(ex.getTimes() - 1);
-					throw ex;
-				}
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("With braces/case/default", "switch('theValue'){\n"
-				+ "\tcase 'notTheValue':\n"
-				+ "\t\tmsg('Nope')\n"
-				+ "\t\tbreak();\n"
-				+ "\tcase 'theValue':\n"
-				+ "\t\tmsg('Success')\n"
-				+ "\t\tbreak();\n"
-				+ "}"),
-				new ExampleScript("With braces/case/default. Note the lack of fallthrough, even without a break(),"
-				+ " except where two cases are directly back to back.",
-				"@a = 5\nswitch(@a){\n"
-				+ "\tcase 1:\n"
-				+ "\tcase 2:\n"
-				+ "\t\tmsg('1 or 2');\n"
-				+ "\tcase 3..4:\n"
-				+ "\t\tmsg('3 or 4');\n"
-				+ "\t\tbreak(); // This is optional, as it would break here anyways, but is recommended.\n"
-				+ "\tcase 5..6:\n"
-				+ "\tcase 8:\n"
-				+ "\t\tmsg('5, 6, or 8')\n"
-				+ "\tdefault:\n"
-				+ "\t\tmsg('Any other value'); # A default is optional\n"
-				+ "}\n"),
-				new ExampleScript("With default condition", "switch('noMatch'){\n"
-				+ "\tcase 'notIt1':\n"
-				+ "\t\tmsg('Nope');\n"
-				+ "\t\tbreak();\n"
-				+ "\tcase 'notIt2':\n"
-				+ "\t\tmsg('Nope');\n"
-				+ "\t\tbreak();\n"
-				+ "\tdefault:\n"
-				+ "\t\tmsg('Success');\n"
-				+ "\t\tbreak();\n"
-				+ "}"),
-				new ExampleScript("With slices", "switch(5){\n"
-				+ "\tcase 1..2:\n"
-				+ "\t\tmsg('First');\n"
-				+ "\t\tbreak();\n"
-				+ "\tcase 3..5:\n"
-				+ "\t\tmsg('Second');\n"
-				+ "\t\tbreak();\n"
-				+ "\tcase 6..8:\n"
-				+ "\t\tmsg('Third');\n"
-				+ "\t\tbreak();\n"
-				+ "}"),
-				new ExampleScript("Functional usage", "switch('theValue',\n"
-				+ "\t'notTheValue',\n"
-				+ "\t\tmsg('Nope'),\n"
-				+ "\t'theValue',\n"
-				+ "\t\tmsg('Success')\n"
-				+ ")"),
-				new ExampleScript("With multiple matches using an array", "switch('string',\n"
-				+ "\tarray('value1', 'value2', 'string'),\n"
-				+ "\t\tmsg('Match'),\n"
-				+ "\t'value3',\n"
-				+ "\t\tmsg('No match')\n"
-				+ ")"),
-				new ExampleScript("With slices in an array", "switch(5,\n"
-				+ "\tarray(1..2, 3..5),\n"
-				+ "\t\tmsg('First'),\n"
-				+ "\t6..8,\n"
-				+ "\t\tmsg('Second')\n"
-				+ ")")};
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			if(children.size() > 1 && children.get(1).getData() instanceof CFunction
-					&& new StringHandling.sconcat().getName().equals(children.get(1).getData().val())) {
-				//This is the brace/case/default usage of switch, probably. We need
-				//to refactor the data into the old switch format.
-				List<ParseTree> newChildren = new ArrayList<>();
-				newChildren.add(children.get(0)); //Initial child
-				List<ParseTree> c = children.get(1).getChildren();
-				List<ParseTree> lastCodeBlock = new ArrayList<>();
-				CArray conditions = new CArray(t);
-				boolean inCase = false;
-				boolean inDefault = false;
-				for(int i = 0; i < c.size(); i++) {
-					//Need up to a 2 lookahead
-					ParseTree c1 = c.get(i);
-					ParseTree c2 = null;
-					if(i + 1 < c.size()) {
-						c2 = c.get(i + 1);
-					}
-					if(CKeyword.isKeyword(c1, "case")) {
-						//If this is a case AND the next one is
-						//a label, this is a case.
-						if(c2 != null && c2.getData() instanceof CLabel) {
-							if(inDefault) {
-								//Default must come last
-								throw new ConfigCompileException("Unexpected case; the default case must come last.", t);
-							}
-							if(lastCodeBlock.size() > 0) {
-								//Ok, need to push some stuff on to the new children
-								newChildren.add(new ParseTree(conditions, c2.getFileOptions()));
-								conditions = new CArray(t);
-								ParseTree codeBlock = new ParseTree(new CFunction(new StringHandling.sconcat().getName(), t), c2.getFileOptions());
-								for(ParseTree line : lastCodeBlock) {
-									codeBlock.addChild(line);
-								}
-								if(codeBlock.getChildren().size() == 1) {
-									codeBlock = codeBlock.getChildAt(0);
-								}
-								newChildren.add(codeBlock);
-								lastCodeBlock = new ArrayList<>();
-							}
-							//Yes, it is. Now we also have to look ahead for
-							//other cases, because
-							//case 1:
-							//case 2:
-							//	code()
-							//would be turned into array(1, 2), code() in the
-							//old style.
-							conditions.push(((CLabel) c2.getData()).cVal(), t);
-							inCase = true;
-							i++;
-							continue;
-						}
-					}
-					if(c1.getData() instanceof CLabel && CKeyword.isKeyword(((CLabel) c1.getData()).cVal(), "default")) {
-						//Default case
-						if(lastCodeBlock.size() > 0) {
-							//Ok, need to push some stuff on to the new children
-							newChildren.add(new ParseTree(conditions, c1.getFileOptions()));
-							conditions = new CArray(t);
-							ParseTree codeBlock = new ParseTree(new CFunction(new StringHandling.sconcat().getName(), t), c1.getFileOptions());
-							for(ParseTree line : lastCodeBlock) {
-								codeBlock.addChild(line);
-							}
-							if(codeBlock.getChildren().size() == 1) {
-								codeBlock = codeBlock.getChildAt(0);
-							}
-							newChildren.add(codeBlock);
-							lastCodeBlock = new ArrayList<>();
-						} else if(conditions.size() > 0) {
-							//Special case where they have
-							//case 0:
-							//default:
-							//	code();
-							//This causes there to be conditions, but no code,
-							//which throws off the argument length. In actuality,
-							//we can simply throw out the conditions, because
-							//this block of code will run if 0 or the default is
-							//hit, and if 0 is the condition provided, it would
-							//work the same if it weren't specified at all.
-							conditions = new CArray(t);
-						}
-						inDefault = true;
-						continue;
-					}
-
-					//Loop forward until we get to the next case
-					if(inCase || inDefault) {
-						lastCodeBlock.add(c1);
-					}
-				}
-				if(conditions.size() > 0) {
-					newChildren.add(new ParseTree(conditions, children.get(0).getFileOptions()));
-				}
-				if(lastCodeBlock.size() > 0) {
-					ParseTree codeBlock = new ParseTree(new CFunction(new StringHandling.sconcat().getName(), t), lastCodeBlock.get(0).getFileOptions());
-					for(ParseTree line : lastCodeBlock) {
-						codeBlock.addChild(line);
-					}
-					if(codeBlock.getChildren().size() == 1) {
-						codeBlock = codeBlock.getChildAt(0);
-					}
-					newChildren.add(codeBlock);
-				}
-				children.clear();
-				children.addAll(newChildren);
-			}
-
-			//Loop through all the conditions and make sure each is unique. Also
-			//make sure that each value is not dynamic.
-			String notConstant = "Cases for a switch statement must be constant, not variable";
-			String alreadyContains = "The switch statement already contains a case for this value, remove the duplicate value";
-			final equals equals = new equals();
-			Set<Mixed> values = new TreeSet<>((Mixed t1, Mixed t2) -> {
-				if(equals.exec(Target.UNKNOWN, null, t1, t2).getBoolean()) {
-					return 0;
-				} else {
-					return t1.val().compareTo(t2.val());
-				}
-			});
-			final boolean hasDefaultCase = (children.size() & 0b00000001) == 0; // size % 2 == 0 -> Even number means there is a default.
-			for(int i = 1; i < children.size(); i += 2) {
-				if(hasDefaultCase && i == children.size() - 1) {
-					// This is the default case code. Stop checking here.
-					break;
-				}
-				//To standardize the rest of the code (and to optimize), go ahead and resolve array()
-				if(children.get(i).getData() instanceof CFunction
-						&& new DataHandling.array().getName().equals(children.get(i).getData().val())) {
-					CArray data = new CArray(t);
-					for(ParseTree child : children.get(i).getChildren()) {
-						if(Construct.IsDynamicHelper(child.getData())) {
-							throw new ConfigCompileException(notConstant, child.getTarget());
-						}
-						data.push(child.getData(), t);
-					}
-					children.set(i, new ParseTree(data, children.get(i).getFileOptions()));
-				}
-				//Now we validate that the values are constant and non-repeating.
-				if(children.get(i).getData().isInstanceOf(CArray.class)) {
-					List<Mixed> list = ((CArray) children.get(i).getData()).asList();
-					for(Mixed c : list) {
-						if(c instanceof CSlice) {
-							for(Mixed cc : ((CSlice) c).asList()) {
-								if(values.contains(cc)) {
-									throw new ConfigCompileException(alreadyContains, cc.getTarget());
-								}
-								values.add(cc);
-							}
-						} else {
-							if(Construct.IsDynamicHelper(c)) {
-								throw new ConfigCompileException(notConstant, c.getTarget());
-							}
-							if(values.contains(c)) {
-								throw new ConfigCompileException(alreadyContains, c.getTarget());
-							}
-							values.add(c);
-						}
-					}
-				} else {
-					Mixed c = children.get(i).getData();
-					if(Construct.IsDynamicHelper(c)) {
-						throw new ConfigCompileException(notConstant, c.getTarget());
-					}
-					if(values.contains(c)) {
-						throw new ConfigCompileException(alreadyContains, c.getTarget());
-					}
-					values.add(c);
-				}
-			}
-
-			if((children.size() > 3 || (children.size() > 1 && children.get(1).getData().isInstanceOf(CArray.class)))
-					//No point in doing this optimization if there are only 3 args and the case is flat.
-					//Also, doing this check prevents an inifinite loop during optimization.
-					&& (children.size() > 0 && !Construct.IsDynamicHelper(children.get(0).getData()))) {
-				ParseTree toReturn = null;
-				//The item passed in is constant (or has otherwise been made constant)
-				//so we can go ahead and condense this down to the single code path
-				//in the switch.
-				for(int i = 1; i < children.size(); i += 2) {
-					Mixed data = children.get(i).getData();
-
-					if(!(data.isInstanceOf(CArray.class)) || data instanceof CSlice) {
-						//Put it in an array to make the rest of this parsing easier.
-						data = new CArray(t);
-						((CArray) data).push(children.get(i).getData(), t);
-					}
-					for(Mixed value : ((CArray) data).asList()) {
-						if(value instanceof CSlice) {
-							long rangeLeft = ((CSlice) value).getStart();
-							long rangeRight = ((CSlice) value).getFinish();
-							if(children.get(0).getData().isInstanceOf(CInt.class)) {
-								long v = Static.getInt(children.get(0).getData(), t);
-								if((rangeLeft < rangeRight && v >= rangeLeft && v <= rangeRight)
-										|| (rangeLeft > rangeRight && v >= rangeRight && v <= rangeLeft)
-										|| (rangeLeft == rangeRight && v == rangeLeft)) {
-									toReturn = children.get(i + 1);
-									break;
-								}
-							}
-						} else if(equals.exec(t, null, children.get(0).getData(), value).getBoolean()) {
-							toReturn = children.get(i + 1);
-							break;
-						}
-					}
-				}
-				//None of the values match. Return the default case, if it exists, or remove the switch entirely
-				//if it doesn't.
-				if(toReturn == null) {
-					if(children.size() % 2 == 0) {
-						toReturn = children.get(children.size() - 1);
-					} else {
-						return Optimizable.REMOVE_ME;
-					}
-				}
-				//Unfortunately, we can't totally remove this, because otherwise break()s in the code
-				//will go unchecked, so we need to keep switch in the code somehow. To make it easy though,
-				//we'll make the most efficient switch we can.
-				ParseTree ret = new ParseTree(new CFunction(new _switch().getName(), t), fileOptions);
-				ret.addChild(new ParseTree(new CInt(1, t), fileOptions));
-				ret.addChild(new ParseTree(new CInt(1, t), fileOptions));
-				ret.addChild(toReturn);
-				return ret;
-			}
-			return null;
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(OptimizationOption.OPTIMIZE_DYNAMIC, OptimizationOption.PRIORITY_OPTIMIZATION);
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> branches = new ArrayList<>(children.size());
-			branches.add(false);
-			if(children.size() == 2) {
-				branches.add(true);
-			} else {
-				for(int i = 1; i < children.size() - 1; i += 2) {
-					branches.add(false);
-					branches.add(true);
-				}
-				if(children.size() % 2 == 0) {
-					branches.add(true);
-				}
-			}
-			return branches;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			// It's the exact same logic as the branches
-			return isBranch(children);
-		}
 	}
 
 	@api
