+++ src/main/java/com/laytonsmith/core/functions/DataHandling.java
@@ -3,10 +3,8 @@ package com.laytonsmith.core.functions;
 import com.laytonsmith.PureUtilities.Version;
 import com.laytonsmith.abstraction.MCCommandSender;
 import com.laytonsmith.annotations.api;
-import com.laytonsmith.annotations.breakable;
 import com.laytonsmith.annotations.core;
 import com.laytonsmith.annotations.hide;
-import com.laytonsmith.annotations.noboilerplate;
 import com.laytonsmith.annotations.nolinking;
 import com.laytonsmith.annotations.noprofile;
 import com.laytonsmith.annotations.seealso;
@@ -25,7 +23,6 @@ import com.laytonsmith.core.Static;
 import com.laytonsmith.core.compiler.BranchStatement;
 import com.laytonsmith.core.compiler.FileOptions;
 import com.laytonsmith.core.compiler.VariableScope;
-import com.laytonsmith.core.compiler.keywords.InKeyword;
 import com.laytonsmith.core.constructs.Auto;
 import com.laytonsmith.core.constructs.CArray;
 import com.laytonsmith.core.constructs.CBoolean;
@@ -36,7 +33,6 @@ import com.laytonsmith.core.constructs.CDouble;
 import com.laytonsmith.core.constructs.CFunction;
 import com.laytonsmith.core.constructs.CIClosure;
 import com.laytonsmith.core.constructs.CInt;
-import com.laytonsmith.core.constructs.CKeyword;
 import com.laytonsmith.core.constructs.CLabel;
 import com.laytonsmith.core.constructs.CMutablePrimitive;
 import com.laytonsmith.core.constructs.CNull;
@@ -57,7 +53,6 @@ import com.laytonsmith.core.exceptions.CRE.CREFormatException;
 import com.laytonsmith.core.exceptions.CRE.CREIllegalArgumentException;
 import com.laytonsmith.core.exceptions.CRE.CREIncludeException;
 import com.laytonsmith.core.exceptions.CRE.CREIndexOverflowException;
-import com.laytonsmith.core.exceptions.CRE.CREInsufficientArgumentsException;
 import com.laytonsmith.core.exceptions.CRE.CREInsufficientPermissionException;
 import com.laytonsmith.core.exceptions.CRE.CREInvalidProcedureException;
 import com.laytonsmith.core.exceptions.CRE.CRERangeException;
@@ -67,21 +62,13 @@ import com.laytonsmith.core.exceptions.CancelCommandException;
 import com.laytonsmith.core.exceptions.ConfigCompileException;
 import com.laytonsmith.core.exceptions.ConfigCompileGroupException;
 import com.laytonsmith.core.exceptions.ConfigRuntimeException;
-import com.laytonsmith.core.exceptions.FunctionReturnException;
-import com.laytonsmith.core.exceptions.LoopBreakException;
-import com.laytonsmith.core.exceptions.LoopContinueException;
 import com.laytonsmith.core.exceptions.StackTraceManager;
-import com.laytonsmith.core.natives.interfaces.Iterator;
 import com.laytonsmith.core.natives.interfaces.Mixed;
 import com.laytonsmith.tools.docgen.templates.ArrayIteration;
-import com.laytonsmith.tools.docgen.templates.Loops;
-
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.HashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.logging.Level;
@@ -426,1113 +413,6 @@ public class DataHandling {
 				new ExampleScript("Operator syntax using combined operators", "@variable = 5;\n@variable -= 10;\nmsg(@variable);"),
 				new ExampleScript("Operator syntax using combined operators", "@variable = 5;\n@variable *= 10;\nmsg(@variable);"),
 				new ExampleScript("Operator syntax using combined operators", "@variable = 5;\n@variable /= 10;\nmsg(@variable);")};
-		}
-	}
-
-	@api
-	@noboilerplate
-	@breakable
-	@seealso({com.laytonsmith.tools.docgen.templates.Loops.class, com.laytonsmith.tools.docgen.templates.ArrayIteration.class})
-	public static class _for extends AbstractFunction implements Optimizable, BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "for";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{4};
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) {
-			return CVoid.VOID;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			return new forelse(true).execs(t, env, parent, nodes);
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CRECastException.class};
-		}
-
-		@Override
-		public String docs() {
-			return "void {assign, condition, expression1, expression2} Acts as a typical for loop. The assignment is first run. Then, a"
-					+ " condition is checked. If that condition is checked and returns true, expression2 is run. After that, expression1 is run. In java"
-					+ " syntax, this would be: for(assign; condition; expression1){expression2}. assign must be an ivariable, either a "
-					+ "pre defined one, or the results of the assign() function. condition must be a boolean.";
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_0_1;
-		}
-		//Doesn't matter, run out of state
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage", "for(assign(@i, 0), @i < 5, @i++,\n\tmsg(@i)\n)"),
-				new ExampleScript("With braces", "for(assign(@i, 0), @i < 2, @i++){\n\tmsg(@i)\n}"),
-				new ExampleScript("With continue. (See continue() for more examples)", "for(assign(@i, 0), @i < 2, @i++){\n"
-				+ "\tif(@i == 1, continue())\n"
-				+ "\tmsg(@i)\n"
-				+ "}")};
-		}
-
-		@Override
-		public LogLevel profileAt() {
-			return LogLevel.WARNING;
-		}
-
-		@Override
-		public String profileMessageS(List<ParseTree> args) {
-			return "Executing function: " + this.getName() + "("
-					+ args.get(0).toStringVerbose() + ", " + args.get(1).toStringVerbose()
-					+ ", " + args.get(2).toStringVerbose() + ", <code>)";
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			//In for(@i = 0, @i < @x, @i++, ...), the @i++ is more optimally written as ++@i, but
-			//it is commonplace to use postfix operations, so if the condition is in fact that simple,
-			//let's reverse it.
-			boolean isInc;
-			try {
-				if(children.get(2).getData() instanceof CFunction
-						&& ((isInc = children.get(2).getData().val().equals("postinc"))
-						|| children.get(2).getData().val().equals("postdec"))
-						&& children.get(2).getChildAt(0).getData() instanceof IVariable) {
-					ParseTree pre = new ParseTree(new CFunction(isInc ? "inc" : "dec", t), children.get(2).getFileOptions());
-					pre.addChild(children.get(2).getChildAt(0));
-					children.set(2, pre);
-				}
-			} catch (IndexOutOfBoundsException e) {
-				//Just ignore it. It's a compile error, but we'll let the rest of the
-				//existing system sort that out.
-			}
-
-			return null;
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(OptimizationOption.OPTIMIZE_DYNAMIC);
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>();
-			ret.add(false);
-			ret.add(false);
-			ret.add(true);
-			ret.add(true);
-			return ret;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>();
-			ret.add(true);
-			ret.add(false);
-			ret.add(false);
-			ret.add(true);
-			return ret;
-		}
-
-	}
-
-	@api
-	@noboilerplate
-	@breakable
-	public static class forelse extends AbstractFunction implements BranchStatement, VariableScope {
-
-		public forelse() {
-		}
-
-		boolean runAsFor = false;
-
-		forelse(boolean runAsFor) {
-			this.runAsFor = runAsFor;
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			return null;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) throws ConfigRuntimeException {
-			ParseTree assign = nodes[0];
-			ParseTree condition = nodes[1];
-			ParseTree expression = nodes[2];
-			ParseTree runnable = nodes[3];
-			ParseTree elseCode = null;
-			if(!runAsFor) {
-				elseCode = nodes[4];
-			}
-			boolean hasRunOnce = false;
-
-			Mixed counter = parent.eval(assign, env);
-			if(!(counter instanceof IVariable)) {
-				throw new CRECastException("First parameter of for must be an ivariable", t);
-			}
-			int _continue = 0;
-			while(true) {
-				boolean cond = ArgumentValidation.getBoolean(parent.seval(condition, env), t);
-				if(cond == false) {
-					break;
-				}
-				hasRunOnce = true;
-				if(_continue >= 1) {
-					--_continue;
-					parent.eval(expression, env);
-					continue;
-				}
-				try {
-					parent.eval(runnable, env);
-				} catch (LoopBreakException e) {
-					int num = e.getTimes();
-					if(num > 1) {
-						e.setTimes(--num);
-						throw e;
-					}
-					return CVoid.VOID;
-				} catch (LoopContinueException e) {
-					_continue = e.getTimes() - 1;
-					parent.eval(expression, env);
-					continue;
-				}
-				parent.eval(expression, env);
-			}
-			if(!hasRunOnce && !runAsFor && elseCode != null) {
-				parent.eval(elseCode, env);
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public String getName() {
-			return "forelse";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{5};
-		}
-
-		@Override
-		public String docs() {
-			return "void {assign, condition, expression1, expression2, else} Works like a normal for loop, but if upon checking the condition the first time,"
-					+ " it is determined that it is false (that is, NO code loops are going to be run) the else code is run instead. If the loop runs,"
-					+ " even once, it will NOT run the else branch. In general, brace syntax and use of for(){ } else { } syntax is preferred, instead"
-					+ " of using forelse directly.";
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_1;
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			return Arrays.asList(false, false, true, true, true);
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			return Arrays.asList(true, false, false, true, true);
-		}
-
-	}
-
-	@api(environments = CommandHelperEnvironment.class)
-	@breakable
-	@seealso({com.laytonsmith.tools.docgen.templates.Loops.class, ArrayIteration.class})
-	public static class foreach extends AbstractFunction implements Optimizable, BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "foreach";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{2, 3, 4};
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws CancelCommandException, ConfigRuntimeException {
-			return CVoid.VOID;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			if(nodes.length < 3) {
-				throw new CREInsufficientArgumentsException("Insufficient arguments passed to " + getName(), t);
-			}
-			ParseTree array = nodes[0];
-			ParseTree key = null;
-			int offset = 0;
-			if(nodes.length == 4) {
-				//Key and value provided
-				key = nodes[1];
-				offset = 1;
-			}
-			ParseTree value = nodes[1 + offset];
-			ParseTree code = nodes[2 + offset];
-			Mixed arr = parent.seval(array, env);
-			Mixed ik = null;
-			if(key != null) {
-				ik = parent.eval(key, env);
-				if(!(ik instanceof IVariable)) {
-					throw new CRECastException("Parameter 2 of " + getName() + " must be an ivariable", t);
-				}
-			}
-			Mixed iv = parent.eval(value, env);
-			if(arr instanceof CSlice) {
-				long start = ((CSlice) arr).getStart();
-				long finish = ((CSlice) arr).getFinish();
-				if(finish < start) {
-					arr = new ArrayHandling.range().exec(t, env, new CInt(start, t), new CInt(finish - 1, t), new CInt(-1, t));
-				} else {
-					arr = new ArrayHandling.range().exec(t, env, new CInt(start, t), new CInt(finish + 1, t));
-				}
-			}
-			if(!(arr instanceof com.laytonsmith.core.natives.interfaces.Iterable)) {
-				throw new CRECastException("Parameter 1 of " + getName() + " must be an Iterable data structure", t);
-			}
-			if(!(iv instanceof IVariable)) {
-				throw new CRECastException("Parameter " + (2 + offset) + " of " + getName() + " must be an ivariable", t);
-			}
-			com.laytonsmith.core.natives.interfaces.Iterable one
-				= (com.laytonsmith.core.natives.interfaces.Iterable) arr;
-			IVariable kkey = (IVariable) ik;
-			IVariable two = (IVariable) iv;
-			if(one.isAssociative()) {
-				//Iteration of an associative array is much easier, and we have
-				//special logic here to decrease the complexity.
-
-				//Clone the set, so changes in the array won't cause changes in
-				//the iteration order.
-				Set<Mixed> keySet = new LinkedHashSet<>(one.keySet());
-				//Continues in an associative array are slightly different, so
-				//we have to track this differently. Basically, we skip the
-				//next element in the array key set.
-				int continues = 0;
-				for(Mixed c : keySet) {
-					if(continues > 0) {
-						//If continues is greater than 0, continue in the loop,
-						//however many times necessary to make it 0.
-						continues--;
-						continue;
-					}
-					//If the key isn't null, set that in the variable table.
-					if(kkey != null) {
-						env.getEnv(GlobalEnv.class).GetVarList().set(new IVariable(kkey.getDefinedType(),
-								kkey.getVariableName(), c, t, env));
-					}
-					//Set the value in the variable table
-					env.getEnv(GlobalEnv.class).GetVarList().set(new IVariable(two.getDefinedType(),
-							two.getVariableName(), one.get(c.val(), t), t, env));
-					try {
-						//Execute the code
-						parent.eval(code, env);
-						//And handle any break/continues.
-					} catch (LoopBreakException e) {
-						int num = e.getTimes();
-						if(num > 1) {
-							e.setTimes(--num);
-							throw e;
-						}
-						return CVoid.VOID;
-					} catch (LoopContinueException e) {
-						// In associative arrays, (unlike with normal arrays) we need to decrement it by one, because the nature of
-						// the normal array is such that the counter is handled manually by our code. Because we are letting java
-						// handle our code though, this run actually counts as one run.
-						continues += e.getTimes() - 1;
-					}
-				}
-				return CVoid.VOID;
-			} else {
-				//It's not associative, so we have more complex handling. We will create an ArrayAccessIterator,
-				//and store that in the environment. As the array is iterated, underlying changes in the array
-				//will be reflected in the object, and we will adjust as necessary. The reason we use this mechanism
-				//is to avoid cloning the array, and iterating that. Arrays may be extremely large, and cloning the
-				//entire array is wasteful in that case. We are essentially tracking deltas this way, which prevents
-				//memory usage from getting out of hand.
-				Iterator iterator = new Iterator(one);
-				List<Iterator> arrayAccessList = env.getEnv(GlobalEnv.class).GetArrayAccessIterators();
-				try {
-					arrayAccessList.add(iterator);
-					int continues = 0;
-					while(true) {
-						int current = iterator.getCurrent();
-						if(continues > 0) {
-							//We have some continues to handle. Blacklisted
-							//values don't count for the continuing count, so
-							//we have to consider that when counting.
-							iterator.incrementCurrent();
-							if(iterator.isBlacklisted(current)) {
-								continue;
-							} else {
-								--continues;
-								continue;
-							}
-						}
-						if(current >= one.size()) {
-							//Done with the iterations.
-							break;
-						}
-						//If the item is blacklisted, we skip it.
-						if(!iterator.isBlacklisted(current)) {
-							if(kkey != null) {
-								env.getEnv(GlobalEnv.class).GetVarList().set(new IVariable(kkey.getDefinedType(),
-										kkey.getVariableName(), new CInt(current, t), t, env));
-							}
-							env.getEnv(GlobalEnv.class).GetVarList().set(new IVariable(two.getDefinedType(),
-									two.getVariableName(), one.get(current, t), t, env));
-							try {
-								parent.eval(code, env);
-							} catch (LoopBreakException e) {
-								int num = e.getTimes();
-								if(num > 1) {
-									e.setTimes(--num);
-									throw e;
-								}
-								return CVoid.VOID;
-							} catch (LoopContinueException e) {
-								continues += e.getTimes();
-								continue;
-							}
-						}
-						iterator.incrementCurrent();
-					}
-				} finally {
-					arrayAccessList.remove(iterator);
-				}
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CRECastException.class, CRERangeException.class};
-		}
-
-		@Override
-		public String docs() {
-			return "void {array, [key], ivar, code} Walks through array, setting ivar equal to each element in the array, then running code."
-					+ " In addition, foreach(1..4, @i, code()) is also valid, setting @i to 1, 2, 3, 4 each time. The same syntax is valid as"
-					+ " in an array slice. If key is set (it must be an ivariable) then the index of each iteration will be set to that."
-					+ " See the examples for a demonstration. ---- "
-					+ " Enhanced syntax may also be used in foreach, using the \"in\", \"as\" and \"else\" keywords. See the examples for"
-					+ " examples of each structure. Using these keywords makes the structure of the foreach read much better. For instance,"
-					+ " with foreach(@value in @array){ } the code very literally reads \"for each value in array\", making ascertaining"
-					+ " the behavior of the loop easier. The \"as\" keyword reads less plainly, and so is not recommended for use, but is"
-					+ " allowed. Note that the array and value are reversed with the \"as\" keyword. An \"else\" block may be used after"
-					+ " the foreach, which will only run if the array provided is empty, that is, the loop code would never run. This provides"
-					+ " a good way to provide \"default\" handling. Array modifications while iterating are supported, and are well defined."
-					+ " See [[CommandHelper/Staged/Array_iteration|the page documenting array iterations]] for full details.";
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_0_1;
-		}
-		//Doesn't matter, runs out of state anyways
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Using \"in\" keyword", "@array = array(1, 2, 3);\n"
-				+ "foreach(@value in @array){\n"
-				+ "\tmsg(@value);\n"
-				+ "}"),
-				new ExampleScript("Using \"in\" keyword, with a key", "@array = array(1, 2, 3);\n"
-				+ "foreach(@key: @value in @array){\n"
-				+ "\tmsg(@key . ': ' . @value);\n"
-				+ "}"),
-				new ExampleScript("Using \"as\" keyword", "@array = array(1, 2, 3);\n"
-				+ "foreach(@array as @value){\n"
-				+ "\tmsg(@value);\n"
-				+ "}"),
-				/*
-		// This is actually borked in real code, so it needs to be fixed. In the meantime, whatever, just remove the example.
-		new ExampleScript("Using \"as\" keyword, with a key", "@array = array(1, 2, 3);\n"
-		+ "foreach(@array as @key: @value){\n"
-		+ "\tmsg(@key . ': ' . @value);\n"
-		+ "}"),*/
-				new ExampleScript("With else clause", "@array = array() # Note empty array\n"
-				+ "foreach(@value in @array){\n"
-				+ "\tmsg(@value);\n"
-				+ "} else {\n"
-				+ "\tmsg('No values were in the array');\n"
-				+ "}"),
-				new ExampleScript("Basic functional usage", "assign(@array, array(1, 2, 3))\nforeach(@array, @i,\n\tmsg(@i)\n)"),
-				new ExampleScript("With braces", "assign(@array, array(1, 2, 3))\nforeach(@array, @i){\n\tmsg(@i)\n}"),
-				new ExampleScript("With a slice", "foreach(1..3, @i){\n\tmsg(@i)\n}"),
-				new ExampleScript("With a slice, counting down", "foreach(3..1, @i){\n\tmsg(@i)\n}"),
-				new ExampleScript("With array keys", "@array = array('one': 1, 'two': 2)\nforeach(@array, @key, @value){\n\tmsg(@key.':'.@value)\n}")};
-		}
-
-		@Override
-		public LogLevel profileAt() {
-			return LogLevel.WARNING;
-		}
-
-		@Override
-		public String profileMessageS(List<ParseTree> args) {
-			return "Executing function: " + this.getName() + "("
-					+ args.get(0).toStringVerbose() + ", " + args.get(1).toStringVerbose()
-					+ ", <code>)";
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(OptimizationOption.OPTIMIZE_DYNAMIC);
-		}
-
-		private static final String CENTRY = new Compiler.centry().getName();
-		private static final String ASSIGN = new assign().getName();
-		private static final String SCONCAT = new StringHandling.sconcat().getName();
-		private static final String IN = new InKeyword().getKeywordName();
-
-		private boolean isFunction(ParseTree node, String function) {
-			return node.getData() instanceof CFunction && node.getData().val().equals(function);
-		}
-
-		private boolean isKeyword(ParseTree node, String keyword) {
-			return node.getData() instanceof CKeyword && node.getData().val().equals(keyword);
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			if(children.size() < 2) {
-				throw new ConfigCompileException("Invalid number of arguments passed to " + getName(), t);
-			}
-			if(isFunction(children.get(0), CENTRY)) {
-				// This is what "@key: @value in @array" looks like initially. We'll refactor this so the next segment can take over properly.
-				ParseTree sconcat = new ParseTree(new CFunction(new StringHandling.sconcat().getName(), t), fileOptions);
-				sconcat.addChild(children.get(0).getChildAt(0));
-				for(int i = 0; i < children.get(0).getChildAt(1).numberOfChildren(); i++) {
-					sconcat.addChild(children.get(0).getChildAt(1).getChildAt(i));
-				}
-				children.set(0, sconcat);
-			}
-			if(children.get(0).getData() instanceof CFunction && children.get(0).getData().val().equals(new StringHandling.sconcat().getName())) {
-				// We may be looking at a "@value in @array" or "@array as @value" type
-				// structure, so we need to re-arrange this into the standard format.
-				ParseTree array = null;
-				ParseTree key = null;
-				ParseTree value = null;
-				List<ParseTree> c = children.get(0).getChildren();
-				if(c.size() == 3) {
-					// No key specified
-					switch(c.get(1).getData().val()) {
-						case "in":
-							// @value in @array
-							value = c.get(0);
-							array = c.get(2);
-							break;
-						case "as":
-							// @array as @value
-							value = c.get(2);
-							array = c.get(0);
-							break;
-					}
-				} else if(c.size() == 4) {
-					if("in".equals(c.get(2).getData().val())) {
-						// @key: @value in @array
-						key = c.get(0);
-						value = c.get(1);
-						array = c.get(3);
-					} else if("as".equals(c.get(1).getData().val())) {
-						// @array as @key: @value
-						array = c.get(0);
-						key = c.get(2);
-						value = c.get(3);
-					}
-				}
-				if(key != null && key.getData() instanceof CLabel) {
-					if(!(((CLabel) key.getData()).cVal() instanceof IVariable)
-							&& !(((CLabel) key.getData()).cVal() instanceof CFunction
-							&& ((CLabel) key.getData()).cVal().val().equals(ASSIGN))) {
-						throw new ConfigCompileException("Expected a variable for key, but \"" + key.getData().val() + "\" was found", t);
-					}
-					key.setData(((CLabel) key.getData()).cVal());
-				}
-				// Now set up the new tree, and return that. Since foreachelse overrides us, we
-				// need to accept all the arguments after the first, and put those in.
-				List<ParseTree> newChildren = new ArrayList<>();
-				newChildren.add(array);
-				if(key != null) {
-					newChildren.add(key);
-				}
-				newChildren.add(value);
-				for(int i = 1; i < children.size(); i++) {
-					newChildren.add(children.get(i));
-				}
-				children.clear();
-				children.addAll(newChildren);
-				// Change foreach(){ ... } else { ... } to a foreachelse.
-				if(children.get(children.size() - 1).getData() instanceof CFunction
-						&& children.get(children.size() - 1).getData().val().equals("else")) {
-					ParseTree foreachelse = new ParseTree(new CFunction(new foreachelse().getName(), t), fileOptions);
-					children.set(children.size() - 1, children.get(children.size() - 1).getChildAt(0));
-					foreachelse.setChildren(children);
-					return foreachelse;
-				}
-			}
-			return null;
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>(children.size());
-			ret.add(false);
-			if(children.size() == 4) {
-				// 3 and 4 arguments are the only actually possible ones here
-				ret.add(false);
-			}
-			ret.add(false);
-			ret.add(true);
-			return ret;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>(children.size());
-			for(ParseTree c : children) {
-				ret.add(true);
-			}
-			return ret;
-		}
-
-	}
-
-	@api
-	@noboilerplate
-	@breakable
-	@seealso({foreach.class, Loops.class, ArrayIteration.class})
-	public static class foreachelse extends foreach {
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			ParseTree array = nodes[0];
-			//The last one
-			ParseTree elseCode = nodes[nodes.length - 1];
-
-			Mixed data = parent.seval(array, env);
-
-			if(!(data.isInstanceOf(CArray.class)) && !(data instanceof CSlice)) {
-				throw new CRECastException(getName() + " expects an array for parameter 1", t);
-			}
-
-			if(((CArray) data).isEmpty()) {
-				parent.eval(elseCode, env);
-			} else {
-				ParseTree pass[] = new ParseTree[nodes.length - 1];
-				System.arraycopy(nodes, 0, pass, 0, nodes.length - 1);
-				nodes[0] = new ParseTree(data, null);
-				return super.execs(t, env, parent, pass);
-			}
-
-			return CVoid.VOID;
-		}
-
-		@Override
-		public String getName() {
-			return "foreachelse";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{4, 5};
-		}
-
-		@Override
-		public String docs() {
-			return "void {array, [key], ivar, code, else} Works like a foreach, except if the array is empty, the else code runs instead. That is, if the code"
-					+ " would not run at all, the else condition would. In general, brace syntax and use of foreach(){ } else { } syntax is preferred, instead"
-					+ " of using foreachelse directly.";
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_1;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage, with the else code not running",
-				"@array = array(1, 2, 3)\n"
-				+ "foreachelse(@array, @val,\n"
-				+ "    msg(@val)\n"
-				+ ", #else \n"
-				+ "    msg('No values in the array')\n"
-				+ ")"),
-				new ExampleScript("Empty array, so else block running",
-				"@array = array()\n"
-				+ "foreachelse(@array, @val,\n"
-				+ "    msg(@val)\n"
-				+ ", #else \n"
-				+ "    msg('No values in the array')\n"
-				+ ")")};
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>(children.size());
-			ret.add(false);
-			if(children.size() == 5) {
-				// 4 and 5 arguments are the only actually possible ones here
-				ret.add(false);
-			}
-			ret.add(false);
-			ret.add(true);
-			ret.add(true);
-			return ret;
-		}
-
-	}
-
-	@api
-	@noboilerplate
-	@breakable
-	@seealso({com.laytonsmith.tools.docgen.templates.Loops.class})
-	public static class _while extends AbstractFunction implements BranchStatement, VariableScope {
-
-		@Override
-		public String getName() {
-			return "while";
-		}
-
-		@Override
-		public String docs() {
-			return "void {condition, [code]} While the condition is true, the code is executed. break and continue work"
-					+ " inside a dowhile, but continuing more than once is pointless, since the loop isn't inherently"
-					+ " keeping track of any counters anyways. Breaking multiple times still works however.";
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_1;
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{1, 2};
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return null;
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			try {
-				while(ArgumentValidation.getBoolean(parent.seval(nodes[0], env), t)) {
-					//We allow while(thing()); to be done. This makes certain
-					//types of coding styles possible.
-					if(nodes.length > 1) {
-						try {
-							parent.seval(nodes[1], env);
-						} catch (LoopContinueException e) {
-							//ok.
-						}
-					}
-				}
-			} catch (LoopBreakException e) {
-				if(e.getTimes() > 1) {
-					throw new LoopBreakException(e.getTimes() - 1, t);
-				}
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			return CNull.NULL;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage", "assign(@i, 5)\nwhile(@i > 0,\n"
-				+ "\tmsg(@i)\n"
-				+ "\t@i--\n"
-				+ ")"),
-				new ExampleScript("With a break", "assign(@i, 0)\nwhile(true,\n"
-				+ "\tmsg(@i)\n"
-				+ "\t@i++\n"
-				+ "\tif(@i > 5, break())\n"
-				+ ")")};
-		}
-
-		@Override
-		public LogLevel profileAt() {
-			return LogLevel.WARNING;
-		}
-
-		@Override
-		public String profileMessageS(List<ParseTree> args) {
-			return "Executing function: " + this.getName() + "("
-					+ args.get(0).toStringVerbose() + ", <code>)";
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>();
-			ret.add(false);
-			ret.add(true);
-			return ret;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			return isBranch(children);
-		}
-
-	}
-
-	@api
-	@noboilerplate
-	@breakable
-	@seealso({com.laytonsmith.tools.docgen.templates.Loops.class})
-	public static class _dowhile extends AbstractFunction implements BranchStatement, VariableScope {
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return null;
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			return CNull.NULL;
-		}
-
-		@Override
-		public String getName() {
-			return "dowhile";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{2};
-		}
-
-		@Override
-		public String docs() {
-			return "void {code, condition} Like while, but always runs the code at least once. The condition is checked"
-					+ " after each run of the code, and if it is true, the code is run again. break and continue work"
-					+ " inside a dowhile, but continuing more than once is pointless, since the loop isn't inherently"
-					+ " keeping track of any counters anyways. Breaking multiple times still works however. In general, using brace"
-					+ " syntax is preferred: do { code(); } while(@condition); instead of using dowhile() directly.";
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_1;
-		}
-
-		@Override
-		public boolean useSpecialExec() {
-			return true;
-		}
-
-		@Override
-		public Mixed execs(Target t, Environment env, Script parent, ParseTree... nodes) {
-			try {
-				do {
-					try {
-						parent.seval(nodes[0], env);
-					} catch (LoopContinueException e) {
-						//ok. No matter how many times it tells us to continue, we're only going to continue once.
-					}
-				} while(ArgumentValidation.getBoolean(parent.seval(nodes[1], env), t));
-			} catch (LoopBreakException e) {
-				if(e.getTimes() > 1) {
-					throw new LoopBreakException(e.getTimes() - 1, t);
-				}
-			}
-			return CVoid.VOID;
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage", "do {\n"
-				+ "\tmsg('This will only run once');\n"
-				+ "} while(false);"),
-				new ExampleScript("Pure functional usage", "dowhile(\n"
-				+ "\tmsg('This will only run once')\n"
-				+ ", #while\n"
-				+ "false)")
-			};
-		}
-
-		@Override
-		public LogLevel profileAt() {
-			return LogLevel.WARNING;
-		}
-
-		@Override
-		public String profileMessageS(List<ParseTree> args) {
-			return "Executing function: " + this.getName() + "(<code>, "
-					+ args.get(1).toStringVerbose() + ")";
-		}
-
-		@Override
-		public List<Boolean> isBranch(List<ParseTree> children) {
-			List<Boolean> ret = new ArrayList<>(2);
-			ret.add(true);
-			ret.add(false);
-			return ret;
-		}
-
-		@Override
-		public List<Boolean> isScope(List<ParseTree> children) {
-			return isBranch(children);
-		}
-	}
-
-	@api
-	public static class _break extends AbstractFunction implements Optimizable {
-
-		@Override
-		public String getName() {
-			return "break";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{0, 1};
-		}
-
-		@Override
-		public String docs() {
-			return "nothing {[int]} Stops the current loop. If int is specified, and is greater than 1, the break travels that many loops up. So, if you had"
-					+ " a loop embedded in a loop, and you wanted to break in both loops, you would call break(2). If this function is called outside a loop"
-					+ " (or the number specified would cause the break to travel up further than any loops are defined), the function will fail. If no"
-					+ " argument is specified, it is the same as calling break(1). This function has special compilation rules. The break number"
-					+ " must not be dynamic, or a compile error will occur. An integer must be hard coded into the function.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CRECastException.class};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_1_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws CancelCommandException, ConfigRuntimeException {
-			int num = 1;
-			if(args.length == 1) {
-				num = Static.getInt32(args[0], t);
-			}
-			throw new LoopBreakException(num, t);
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage", "for(assign(@i, 0), @i < 1000, @i++,\n"
-				+ "\tfor(assign(@j, 0), @j < 1000, @j++,\n"
-				+ "\t\tmsg('This will only display once')\n"
-				+ "\t\tbreak(2)\n"
-				+ "\t)"
-				+ ")"),
-				new ExampleScript("Invalid number", "for(assign(@i, 0), @i < 1000, @i++,\n"
-				+ "\tfor(assign(@j, 0), @j < 1000, @j++,\n"
-				+ "\t\tbreak(3) #There are only 2 loops to break out of\n"
-				+ "\t)"
-				+ ")", true)};
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			if(children.size() == 1) {
-				if(children.get(0).isDynamic()) {
-					//This is absolutely a bad design, if there is a variable here
-					//in the break. Due to optimization, this is a compile error.
-					throw new ConfigCompileException("The parameter sent to break() should"
-							+ " be hard coded, and should not be dynamically determinable, since this is always a sign"
-							+ " of loose code flow, which should be avoided.", t);
-				}
-				if(!(children.get(0).getData().isInstanceOf(CInt.class))) {
-					throw new ConfigCompileException("break() only accepts integer values.", t);
-				}
-			}
-			return null;
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(OptimizationOption.OPTIMIZE_DYNAMIC
-			//, OptimizationOption.TERMINAL This can't be added yet, because of things like switch, where code
-			//branches aren't considered correctly.
-			);
-		}
-
-	}
-
-	@api
-	public static class _continue extends AbstractFunction {
-
-		@Override
-		public String getName() {
-			return "continue";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{0, 1};
-		}
-
-		@Override
-		public String docs() {
-			return "void {[int]} Skips the rest of the code in this loop, and starts the loop over, with it continuing at the next index. If this function"
-					+ " is called outside of a loop, the command will fail. If int is set, it will skip 'int' repetitions. If no argument is specified,"
-					+ " 1 is used.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CRECastException.class};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_1_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws CancelCommandException, ConfigRuntimeException {
-			int num = 1;
-			if(args.length == 1) {
-				num = Static.getInt32(args[0], t);
-			}
-			throw new LoopContinueException(num, t);
-		}
-
-		@Override
-		public ExampleScript[] examples() throws ConfigCompileException {
-			return new ExampleScript[]{
-				new ExampleScript("Basic usage", "for(assign(@i, 0), @i < 5, @i++){\n"
-				+ "\tif(@i == 2, continue())\n"
-				+ "\tmsg(@i)\n"
-				+ "}"),
-				new ExampleScript("Argument specified", "for(assign(@i, 0), @i < 5, @i++){\n"
-				+ "\tif(@i == 2, continue(2))\n"
-				+ "\tmsg(@i)\n"
-				+ "}")};
 		}
 	}
 
@@ -2445,58 +1325,6 @@ public class DataHandling {
 	}
 
 	@api
-	public static class _return extends AbstractFunction implements Optimizable {
-
-		@Override
-		public String getName() {
-			return "return";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{0, 1};
-		}
-
-		@Override
-		public String docs() {
-			return "nothing {mixed} Returns the specified value from this procedure. It cannot be called outside a procedure.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return null;
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return false;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_2_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(
-					OptimizationOption.TERMINAL
-			);
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws ConfigRuntimeException {
-			Mixed ret = (args.length == 1 ? args[0] : CVoid.VOID);
-			throw new FunctionReturnException(ret, t);
-		}
-	}
-
-	@api
 	public static class include extends AbstractFunction /*implements Optimizable*/ {
 		// Can't currently optimize this, because it depends on knowing whether or not
 		// we are in cmdline mode, which is included with the environment, which doesn't
@@ -2603,139 +1431,6 @@ public class DataHandling {
 				m += "<dynamic input>";
 			}
 			return m + ")";
-		}
-
-	}
-
-	@api
-	public static class call_proc extends AbstractFunction implements Optimizable {
-
-		@Override
-		public String getName() {
-			return "call_proc";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{Integer.MAX_VALUE};
-		}
-
-		@Override
-		public String docs() {
-			return "mixed {proc_name, [var1...]} Dynamically calls a user defined procedure. call_proc(_myProc, 'var1') is the equivalent of"
-					+ " _myProc('var1'), except you could dynamically build the procedure name if need be. This is useful for dynamic coding,"
-					+ " however, closures work best for callbacks. Throws an InvalidProcedureException if the procedure isn't defined. If you are"
-					+ " hardcoding the first parameter, a warning will be issued, because it is much more efficient and safe to directly use"
-					+ " a procedure if you know what its name is beforehand.";
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CREInvalidProcedureException.class};
-		}
-
-		@Override
-		public boolean isRestricted() {
-			return true;
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_2_0;
-		}
-
-		@Override
-		public Boolean runAsync() {
-			return null;
-		}
-
-		@Override
-		public Mixed exec(Target t, Environment env, Mixed... args) throws ConfigRuntimeException {
-			if(args.length < 1) {
-				throw new CREInsufficientArgumentsException("Expecting at least one argument to " + getName(), t);
-			}
-			Procedure proc = env.getEnv(GlobalEnv.class).GetProcs().get(args[0].val());
-			if(proc != null) {
-				List<Mixed> vars = new ArrayList<>(Arrays.asList(args));
-				vars.remove(0);
-				Environment newEnv = null;
-				try {
-					newEnv = env.clone();
-				} catch (CloneNotSupportedException ex) {
-					throw new RuntimeException(ex);
-				}
-				return proc.execute(vars, newEnv, t);
-			}
-			throw new CREInvalidProcedureException("Unknown procedure \"" + args[0].val() + "\"", t);
-		}
-
-		@Override
-		public Set<OptimizationOption> optimizationOptions() {
-			return EnumSet.of(OptimizationOption.OPTIMIZE_DYNAMIC);
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			if(children.size() < 1) {
-				throw new CREInsufficientArgumentsException("Expecting at least one argument to " + getName(), t);
-			}
-			if(children.get(0).isConst()) {
-				MSLog.GetLogger().Log(MSLog.Tags.COMPILER, LogLevel.WARNING, "Hardcoding procedure name in " + getName() + ", which is inefficient."
-						+ " Consider calling the procedure directly if the procedure name is known at compile time.", t);
-			}
-			return null;
-		}
-
-	}
-
-	@api
-	public static class call_proc_array extends call_proc {
-
-		@Override
-		public Mixed exec(Target t, Environment environment, Mixed... args) throws ConfigRuntimeException {
-			CArray ca = Static.getArray(args[1], t);
-			if(ca.inAssociativeMode()) {
-				throw new CRECastException("Expected the array passed to " + getName() + " to be non-associative.", t);
-			}
-			Mixed[] args2 = new Mixed[(int) ca.size() + 1];
-			args2[0] = args[0];
-			for(int i = 1; i < args2.length; i++) {
-				args2[i] = ca.get(i - 1, t);
-			}
-			return super.exec(t, environment, args2);
-		}
-
-		@Override
-		public Class<? extends CREThrowable>[] thrown() {
-			return new Class[]{CREInvalidProcedureException.class, CRECastException.class};
-		}
-
-		@Override
-		public String getName() {
-			return "call_proc_array";
-		}
-
-		@Override
-		public Integer[] numArgs() {
-			return new Integer[]{2};
-		}
-
-		@Override
-		public String docs() {
-			return "mixed {proc_name, array} Works like call_proc, but allows for variable or unknown number of arguments to be passed to"
-					+ " a proc. The array parameter is \"flattened\", and call_proc is essentially called. If the array is associative, an"
-					+ " exception is thrown.";
-		}
-
-		@Override
-		public MSVersion since() {
-			return MSVersion.V3_3_1;
-		}
-
-		@Override
-		public ParseTree optimizeDynamic(Target t, Environment env, List<ParseTree> children, FileOptions fileOptions) throws ConfigCompileException, ConfigRuntimeException {
-			//If they hardcode the name, that's fine, because the variables may just be the only thing that's variable.
-			return null;
 		}
 
 	}
