+++ src/test/java/com/laytonsmith/core/functions/DataHandlingTest.java
@@ -6,7 +6,6 @@ import com.laytonsmith.abstraction.MCServer;
 import com.laytonsmith.core.MethodScriptCompiler;
 import com.laytonsmith.core.Static;
 import com.laytonsmith.core.environments.CommandHelperEnvironment;
-import com.laytonsmith.core.exceptions.ConfigRuntimeException;
 import com.laytonsmith.testing.StaticTest;
 import static com.laytonsmith.testing.StaticTest.RunCommand;
 import static com.laytonsmith.testing.StaticTest.SRun;
@@ -57,120 +56,6 @@ public class DataHandlingTest {
 	}
 
 	@Test(timeout = 10000)
-	public void testFor1() throws Exception {
-		String config = "/for = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 5), inc(@i),\n"
-				+ "     array_push(@array, @i)\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/for");
-		verify(fakePlayer).sendMessage("{0, 1, 2, 3, 4}");
-	}
-
-	@Test(expected = ConfigRuntimeException.class, timeout = 10000)
-	public void testFor3() throws Exception {
-		String script
-				= "   assign(@array, array())"
-				+ " for('nope', lt(@i, 5), inc(@i),\n"
-				+ "     array_push(@array, @i)\n"
-				+ " )\n"
-				+ " msg(@array)\n";
-		MethodScriptCompiler.execute(MethodScriptCompiler.compile(MethodScriptCompiler.lex(script, null, true), null), env, null, null);
-
-	}
-
-	@Test(timeout = 10000)
-	public void testForeach1() throws Exception {
-		String config = "/for = >>>\n"
-				+ " assign(@array, array(1, 2, 3, 4, 5))\n"
-				+ " assign(@array2, array())"
-				+ " foreach(@array, @i,\n"
-				+ "     array_push(@array2, @i)\n"
-				+ " )\n"
-				+ " msg(@array2)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/for");
-		verify(fakePlayer).sendMessage("{1, 2, 3, 4, 5}");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeach2() throws Exception {
-		String config = "/for = >>>\n"
-				+ " assign(@array, array(1, 2, 3, 4, 5))\n"
-				+ " assign(@array2, array())"
-				+ " foreach(@array, @i,\n"
-				+ "     if(equals(@i, 1), continue(2))"
-				+ "     array_push(@array2, @i)\n"
-				+ " )\n"
-				+ " msg(@array2)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/for");
-		verify(fakePlayer).sendMessage("{3, 4, 5}");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeach3() throws Exception {
-		String config = "/for = >>>\n"
-				+ " assign(@array, array(1, 2, 3, 4, 5))\n"
-				+ " assign(@array1, array(1, 2, 3, 4, 5))\n"
-				+ " assign(@array2, array())\n"
-				+ " foreach(@array1, @j,"
-				+ "     foreach(@array, @i,\n"
-				+ "         if(equals(@i, 3), break(2))"
-				+ "         array_push(@array2, @i)\n"
-				+ "     )\n"
-				+ " )"
-				+ " msg(@array2)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/for");
-		verify(fakePlayer).sendMessage("{1, 2}");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeachWithArraySlice() throws Exception {
-		SRun("foreach(1..2, @i, msg(@i))", fakePlayer);
-		verify(fakePlayer).sendMessage("1");
-		verify(fakePlayer).sendMessage("2");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeachWithKeys1() throws Exception {
-		SRun("@array = array(1: 'one', 2: 'two') @string = '' foreach(@array, @key, @value, @string .= (@key.':'.@value.';')) msg(@string)", fakePlayer);
-		verify(fakePlayer).sendMessage("1:one;2:two;");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeachWithKeys2() throws Exception {
-		SRun("@array = array('one': 1, 'two': 2) @string = '' foreach(@array, @key, @value, @string .= (@key.':'.@value.';')) msg(@string)", fakePlayer);
-		verify(fakePlayer).sendMessage("one:1;two:2;");
-	}
-
-	@Test(timeout = 10000)
-	public void testForeachWithKeys3() throws Exception {
-		SRun("@array = array('one': 1, 'two': 2)\nforeach(@array, @key, @value){\n\tmsg(@key.':'.@value)\n}", fakePlayer);
-		verify(fakePlayer).sendMessage("one:1");
-		verify(fakePlayer).sendMessage("two:2");
-	}
-
-	@Test
-	public void testForelse() throws Exception {
-		SRun("forelse(assign(@i, 0), @i < 0, @i++, msg('fail'), msg('pass'))", fakePlayer);
-		verify(fakePlayer).sendMessage("pass");
-		verify(fakePlayer, times(0)).sendMessage("fail");
-	}
-
-	@Test
-	public void testForeachelse() throws Exception {
-		SRun("foreachelse(array(), @val, msg('fail'), msg('pass'))", fakePlayer);
-		SRun("foreachelse(array(1), @val, msg('pass'), msg('fail'))", fakePlayer);
-		SRun("foreachelse(1..2, @val, msg('pass'), msg('fail'))", fakePlayer);
-		verify(fakePlayer, times(4)).sendMessage("pass");
-		verify(fakePlayer, times(0)).sendMessage("fail");
-	}
-
-	@Test(timeout = 10000)
 	public void testCallProcIsProc() throws Exception {
 		when(fakePlayer.isOp()).thenReturn(true);
 		String config = "/for = >>>\n"
@@ -185,87 +70,6 @@ public class DataHandlingTest {
 		verify(fakePlayer).sendMessage("false");
 		verify(fakePlayer).sendMessage("true");
 		verify(fakePlayer).sendMessage("hello world");
-	}
-
-	/**
-	 * There is a bug that causes an infinite loop, so we put a 10 second timeout
-	 *
-	 * @throws Exception
-	 */
-	@Test(timeout = 10000)
-	public void testContinue1() throws Exception {
-		String config = "/continue = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 5), inc(@i),\n"
-				+ "     if(equals(@i, 2), continue(1))\n"
-				+ "     array_push(@array, @i)\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/continue");
-		verify(fakePlayer).sendMessage("{0, 1, 3, 4}");
-	}
-
-	@Test(timeout = 10000)
-	public void testContinue2() throws Exception {
-		String config = "/continue = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 5), inc(@i),\n"
-				+ "     if(equals(@i, 2), continue(2))\n"
-				+ "     array_push(@array, @i)\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/continue");
-		verify(fakePlayer).sendMessage("{0, 1, 4}");
-	}
-
-	@Test(timeout = 10000)
-	public void testContinue3() throws Exception {
-		String config = "/continue = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 5), inc(@i),\n"
-				+ "     if(equals(@i, 2), continue(3))\n"
-				+ "     array_push(@array, @i)\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/continue");
-		verify(fakePlayer).sendMessage("{0, 1}");
-	}
-
-	@Test(timeout = 10000)
-	public void testBreak1() throws Exception {
-		String config = "/break = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 2), inc(@i),\n"
-				+ "     for(assign(@j, 0), lt(@j, 5), inc(@j),\n"
-				+ "         if(equals(@j, 2), break())\n"
-				+ "         array_push(@array, concat('j:', @j))\n"
-				+ "     )\n"
-				+ "     array_push(@array, concat('i:', @i))\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/break");
-		verify(fakePlayer).sendMessage("{j:0, j:1, i:0, j:0, j:1, i:1}");
-	}
-
-	@Test(timeout = 10000)
-	public void testBreak2() throws Exception {
-		String config = "/break = >>>\n"
-				+ " assign(@array, array())"
-				+ " for(assign(@i, 0), lt(@i, 2), inc(@i),\n"
-				+ "     for(assign(@j, 0), lt(@j, 5), inc(@j),\n"
-				+ "         if(equals(@j, 2), break(2))\n"
-				+ "         array_push(@array, concat('j:', @j))\n"
-				+ "     )\n"
-				+ "     array_push(@array, concat('i:', @i))\n"
-				+ " )\n"
-				+ " msg(@array)\n"
-				+ "<<<\n";
-		RunCommand(config, fakePlayer, "/break");
-		verify(fakePlayer).sendMessage("{j:0, j:1}");
 	}
 
 	@Test(timeout = 10000)
@@ -485,18 +289,6 @@ public class DataHandlingTest {
 		verify(fakePlayer2).sendMessage("newlabel");
 	}
 
-	@Test(timeout = 10000)
-	public void testWhile() throws Exception {
-		SRun("assign(@i, 2) while(@i > 0, @i-- msg('hi'))", fakePlayer);
-		verify(fakePlayer, times(2)).sendMessage("hi");
-	}
-
-	@Test(timeout = 10000)
-	public void testDoWhile() throws Exception {
-		SRun("assign(@i, 2) dowhile(@i-- msg('hi'), @i > 0)", fakePlayer);
-		verify(fakePlayer, times(2)).sendMessage("hi");
-	}
-
 	@Test
 	public void testToRadix() throws Exception {
 		assertEquals("f", SRun("to_radix(15, 16)", null));
@@ -513,7 +305,6 @@ public class DataHandlingTest {
 	public void testClosureReturnsFromExecute() throws Exception {
 		assertEquals("3", SRun("execute(closure(return(3)))", fakePlayer));
 	}
-
 
 	@Test
 	public void testEmptyClosureFunction() throws Exception {
